@@ -62,7 +62,7 @@ public class SecurityConfig {
                                 "/api/signup",
                                 "/api/auth/refresh",
                                 "/api/roles",
-                                "/api/utilisateur",
+                                "/api/utilisateur/**",
                                 "/api/trajets/**"
                         ).permitAll()
                         .anyRequest().authenticated());
